A snapshot after big changes
    - change assign.rs to assign module
    - drop dynamic var decay control
    - revert to LR + RSR from EVSIDS
    - use stabilization mode to select var rewarding mode
    - less eliminator invocation
    - cargo update

	modified:   Cargo.lock
	modified:   Cargo.toml
	deleted:    src/assign.rs
	new file:   src/assign/heap.rs
	new file:   src/assign/mod.rs
	new file:   src/assign/propagate.rs
	new file:   src/assign/reward.rs
	new file:   src/assign/select.rs
	new file:   src/assign/stack.rs
	new file:   src/assign/var.rs
	modified:   src/config.rs
	modified:   src/restart.rs
	modified:   src/solver.rs
	modified:   src/state.rs
	modified:   src/validate.rs
